deploy/wechat-bot/requirements.txt
flask==3.0.0
anthropic==0.25.0
requests==2.31.0
pycryptodome==3.20.0
python-dotenv==1.0.0
gunicorn==21.2.0
